% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Wheat APSIM model data set}
\format{dataframe}
\usage{
data(Data)
}
\description{
Wheat APSIM model simulated database
}
\examples{
data(Data)
\donttest{boxplot(Yield~Genotype,data=Data,
col='lightblue',border='darkblue')}
}
\references{
Casadebaig P, Zheng B, Chapman S, Huth N, Faivre R, et al. (2016) Plosone 11(1), e0146385
(\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0146385}{Plosone})
}
\keyword{datasets}
