% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data preprocessing}
\usage{
data_prep(data, trait, Genotype, Environment)
}
\arguments{
\item{data}{a dataframe containing at least three column of trait of interest,
genotype, and environment information.}

\item{trait}{name of the interested column in the dataframe to analysize.}

\item{Genotype}{genetic varieties}

\item{Environment}{environmental parameters}
}
\value{
a dataframe with descriptive statics
}
\description{
\code{data_prep} calculate genetic and environmental basic statstics required
for stability indices.
}
\details{
This function calculate mean, standard deviation (sd),correlation coefficient (cv),
log of variance (logvar), log of mean(logmean) for each genotype. \cr
Mean and maximum value were calculated for each environment.
}
\examples{
data(Data)
result <- data_prep(Data,'Yield','Genotype','Environment')
str(result)#dataframe
}
\author{
Tien Cheng Wang (\email{tien.wang@gem.uni-hannover.de})
}
