% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_val.R
\name{eco_val}
\alias{eco_val}
\title{Ecovalance}
\usage{
eco_val(data, trait, Genotype, Environment)
}
\arguments{
\item{data}{a dataframe containing at least three column of trait of interest,
genotype, and environment information.}

\item{trait}{the interested variable to be analysize.}

\item{Genotype}{a variable labeling different genotypic varieties}

\item{Environment}{a variable lableing different environmental parameters}
}
\value{
a dataframe with environmental variance indices
}
\description{
\code{eco_val} calculate genetic and environmental interaction.
}
\examples{
data(Data)
eco_valance <- eco_val(Data,'Yield','Genotype','Environment')
}
\references{
\insertRef{wricke1962}{toolStability}
}
\author{
Tien Cheng Wang (\email{tien.wang@gem.uni-hannover.de})
}
