% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_var.R
\name{env_var}
\alias{env_var}
\title{Environmental variance}
\usage{
env_var(data, trait, Genotype, Environment)
}
\arguments{
\item{data}{a dataframe containing at least three column of trait of interest,
genotype, and environment information.}

\item{trait}{the interested variable to be analysize.}

\item{Genotype}{a variable labeling different genotypic varieties}

\item{Environment}{a variable lableing different environmental parameters}
}
\value{
a dataframe with environmental variance indices
}
\description{
\code{env_var} calculate variance of a genotype across environments.
}
\details{
This function calculate environmental variance of genotypes with static concept. By detecting all
deviation from the genotypic mean. \cr The larger he environmental variance
one genotype is, the lower the stability.
\deqn{S^{2}_{xi} = \frac{\Simga_{j}(X_{ij} -bar(X)_{i.})^2}{E-1} }
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{bar(X)_{i.}} denoting marginal means of genotype i.
}
\examples{
data(Data)
env_variance <- env_var(Data,'Yield','Genotype','Environment')
}
\references{
\insertRef{roemer1917}{toolStability}
}
\author{
Tien Cheng Wang (\email{tien.wang@gem.uni-hannover.de})
}
