% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_var.R
\name{environmental_variance}
\alias{environmental_variance}
\title{Environmental variance}
\usage{
environmental_variance(trait, genotype, ...)
}
\arguments{
\item{trait}{numeric vector of interested trait to be analysized.}

\item{genotype}{a character or factor vector labeling different genotypic varieties}

\item{...}{data table with precalculated statistic procuded by \code{\link{stability_indices_table}}of to be passed}
}
\value{
a data table with environmental variance indices
}
\description{
\code{env_var} is used to calculate variance of a genotype across environments.
}
\details{
Environmental variance (Roemer 1917) is calculated based on suming up all deviation from genotypic mean for each genotype.
The larger the environmental variance of one genotype is, the lower the stability.

\deqn{S^{2}_{xi} = \frac{\sum_{j}(X_{ij} - \bar(X)_{i.})^2}{E-1} }
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{bar(X)_{i.}} denoting marginal means of genotype i.
}
\examples{
data(Data)
environmental_variance <- env_var(Data$'Yield',Data$Genotype,Data$environment)
}
\references{
\insertRef{roemer1917}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{stability}
\keyword{static}
