% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypic_stability.R
\name{genotypic_stability}
\alias{genotypic_stability}
\title{Genotypic stability}
\usage{
genotypic_stability(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with genotypic stability
}
\description{
\code{genotypic_stability} calculate variance of a genotype across environments.
}
\details{
Genotypic stability (Hanson, 1970) is calculatd based on regression function.
Variety with low stability variance is considered as stable.
Under the linear model
\deqn{Y =\mu + b_{i}e_{j} + g_{i} + d_{ij}}
where Y is the predicted phenotypic values, \eqn{g_{i}}, \eqn{e_{j}} and \eqn{\mu} denoting
genotypic, environmental and overall popluation mean,respectively.

The effect of GE-interaction may be expressed as:
\deqn{(ge)_{ij} = b_{i}e_{j} + d_{ij}}
where \eqn{b_{i}} is the coefficient of regression and \eqn{d_{ij}} a deviation.

Genotypic stability:
\deqn{ D_{i}^{2} = \sum_{j} (X_{ij} - \bar{X_{i.}}- b_{min}\bar{X_{.j}} +  b_{min}\bar{X_{..}}^{2}}

where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{\bar{X_{i.}}} and  \eqn{\bar{X_{.j}}} \cr
denoting marginal means of genotype i and environemnt j,respectively. \cr
\eqn{\bar{X_{..}}} denote the overall mean of X.

\eqn{b_{min}} is the minimum value of (\code{\link{coefficient_of_regression}}) over all enrionment.
}
\examples{
data(Data)
genotypic.stability <- genotypic_stability(Data, "Yield", "Genotype", "Environment")
}
\references{
\insertRef{hanson1970}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{approach}
\keyword{regression}
