% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safty_first_index.R
\name{safty_first_index}
\alias{safty_first_index}
\title{Safty-first Index}
\usage{
safty_first_index(data, trait, genotype, environment, lambda)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}

\item{lambda}{threshold value of trait that define stability.}
}
\value{
a data table with coefficient of determination
}
\description{
\code{safty_first_index} calculate variance of a genotype across environments.
}
\details{
Safty-first index (Eskridge, 1990) is calculatd based on the assumption of
that the trait from each genotype follow normal distribution over enviornments.
Among different environments, trait below a given cirtical level \eqn{\lambda}
is defined as failure of trait. The probability of trait failure can be obtained
by entering mean and variance of trait plus \eqn{\lambda}
into the cumulated density function of normal distribution.
Variety with low safty first index is considered as stable.

\deqn{Pr(Y_{ij} < \lambda) = \Phi \left[
(\lambda - \mu_{i})/ \sqrt \sigma_{ii}
\right]}

where \eqn{\lambda} is the critical subsistence level of trait.
\eqn{\Phi} is the cumulative distribution function of the standard normal distribution. \eqn{\mu_{i}} and \eqn{\sigma_{ii}}
is the mean and varance of the system i.
}
\examples{
data(Data)
safty.first.index <- safty_first_index(Data, "Yield", "Genotype", "Environment", median(Data$Yield))
}
\references{
\insertRef{eskridge1990}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{approach}
\keyword{probabilistic}
