% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_var.R
\name{stab_var}
\alias{stab_var}
\title{Stability variance}
\usage{
stab_var(data, trait, Genotype, Environment)
}
\arguments{
\item{data}{a dataframe containing at least three column of trait of interest,
genotype, and environment information.}

\item{trait}{the interested variable to be analysize.}

\item{Genotype}{a variable labeling different genotypic varieties}

\item{Environment}{a variable lableing different environmental parameters}
}
\value{
a dataframe with stability variance indices
}
\description{
\code{stab_var} calculate variance of a genotype across environments.
}
\details{
This function calculate the unbiased estimation of genotypic and environmental \cr
interaction. The stability variance is a linear combination of the ecovalence.

\deqn{\sigma^{2}_{i} = \frac{1}{(G-1)(G-2)(E-1)}\big[
\Simga_{j}(X_{ij} -bar(X_{i.})_{i.})^2-\Simga_{i}\Simga_{j}(X_{ij} -bar(X_{i.})_{i.})^2
\big]}
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{bar(X)_{i.}} and  \eqn{bar(X)_{.j}} \cr
denoting marginal means of genotype i and environemnt j,respectively. \cr
\eqn{bar(X)_{..}} denote the overall mean of X.
}
\examples{
data(Data)
stab_variance <- stab_var(Data,'Yield','Genotype','Environment')

}
\references{
Shukla, G. K. "Some statistical aspects of partitioning genotype environmental components of variability."
 \emph{Heredity} 29.2 (1972): 237-245.
}
\author{
Tien Cheng Wang (\email{tien.wang@gem.uni-hannover.de})
}
