% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_stability.R
\name{table_stability}
\alias{table_stability}
\title{Table stability}
\usage{
table_stability(data, trait, genotype, environment, lambda,
  normalize = FALSE)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}

\item{lambda}{threshold value of trait that define stability for a genotype across environments through probabilistic approach.}

\item{normalize}{a logical value indicating whether stability indicies should be normalized to the range from 0 to 1, where 1 refer to stable and 0 is unstable. Default is \code{FALSE}.}
}
\value{
a data table with multiple stability indices
}
\description{
\code{table_stability} export all the stability indicies in the package.
}
\details{
Combine all stability indices in this package and export as a table, including mean trait, normality of the trait across environment as well.
}
\examples{
data(Data)
tb <- table_stability(Data,"Yield","Genotype","Environment",median(Data$Yield),normalize = TRUE)
}
\references{
\insertRef{doering2018}{toolStability}
\insertRef{pinthus1973}{toolStability}
\insertRef{finlay1963}{toolStability}
\insertRef{eberhart1966}{toolStability}
\insertRef{wricke1962}{toolStability}
\insertRef{roemer1917}{toolStability}
\insertRef{hanson1970}{toolStability}
\insertRef{lin1988}{toolStability}
\insertRef{shukla1972}{toolStability}
\insertRef{nassar1987}{toolStability}
\insertRef{eskridge1990}{toolStability}
}
\seealso{
\code{\link{adjusted_coefficient_of_variation}}

\code{\link{coefficient_of_determination}}

\code{\link{coefficient_of_regression}}

\code{\link{deviation_mean_squares}}

\code{\link{ecovalence}}

\code{\link{environmental_variance}}

\code{\link{genotypic_stability}}

\code{\link{genotypic_superiority_measure}}

\code{\link{stability_variance}}

\code{\link{variance_of_rank}}

\code{\link{safty_first_index}}
}
\author{
Tien Cheng Wang
}
\keyword{and}
\keyword{approach}
\keyword{dynamic,}
\keyword{nonparametric}
\keyword{probablistic}
\keyword{regression,}
\keyword{static,}
