% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolStability-package.R
\docType{package}
\name{toolStability-package}
\alias{toolStability}
\alias{toolStability-package}
\title{toolStability: Tool for Stability Indices calculation}
\description{
Tools to calculate stability indices with parametric,
 non-parametric and probabilistic approaches. The basic data format requirement for 'toolStability' is a data frame with 3 columns including numeric trait values, genotype,and environmental labels. Output is the dataframe with stability indices for each genotype. It returns the table of genotype and stability indices. "table_stability" offers the summary table of all stability indices in this package.
}
\author{
\strong{Maintainer}: Tien-Cheng Wang \email{tien.wang@gem.uni-hannover.de}

Authors:
\itemize{
  \item Tsu-Wei Chen \email{chen@gem.uni-hannover.de}
}

}
\keyword{internal}
