% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_of_rank.R
\name{variance_of_rank}
\alias{variance_of_rank}
\title{variance of rank}
\usage{
variance_of_rank(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with variance of rank
}
\description{
\code{variance_of_rank} calculate variance of a genotype across environments.
}
\details{
Variance of rank is calculatd based on regression function.
Variety with low variance of rank is considered as stable.

Correction for each genotype i was done by subtraction of marginal genotypic mean \eqn{\bar{X_{i.}}} and the addition of overall mean\eqn{\bar{X_{..}}}.
\deqn{X_{corrected ij} = X_{ij} - \bar{X_{i.}} + \bar{X_{..}}}
Then calculated the rank all genotypes for each environment j
\deqn{r_{ij} = rank(X_{corrected ij})}

Variance of rank is calculated as the following equation.
\deqn{S_{i}4 = \frac{\sum_{j} (r_{ij}-\bar{r_{i.}}^{2}}{E-1}}
where \eqn{r_{ij}} is the rank of genotype i in environment j and \eqn{\bar{r_{i.}}} is the marginal rank of genotype i over environment,
based on the corrected \eqn{X_{ij}} values.
}
\examples{
data(Data)
variance.of.rank <- variance_of_rank(Data,'Yield','Genotype','Environment')

}
\references{
\insertRef{nassar1987}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{approach}
\keyword{nonparametric}
