% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_of_determination.R
\name{coefficient_of_determination}
\alias{coefficient_of_determination}
\title{Coefficient of determination}
\usage{
coefficient_of_determination(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with coefficient of determination
}
\description{
\code{coefficient_of_determination} calculate variance of a genotype across environments.
}
\details{
Coefficient of determination (Pinthus, 1976) is calculatd based on regression function.
Variety with low coefficient of determination is considered as stable.
Under the linear model
\deqn{Y =\mu + b_{i}\cdot e_{j} + g_{i} + d_{ij}}
where Y is the predicted phenotypic values, \eqn{g_{i}}, \eqn{e_{j}} and \eqn{\mu} denoting
genotypic, environmental and overall popluation mean,respectively. \eqn{b_{i}} is the coefficient of regression and \eqn{d_{ij}} denotes deviation.

The effect of GE-interaction may be expressed as:
\deqn{(ge)_{ij} = b_{i}e_{j} + d_{ij}}

Deviation mean squares(\code{\link{deviation_mean_squares}}):
\deqn{
s^{2}_{di} = \frac{1}{E-2} \left [
\sum_{j}(X_{ij} -\bar{X_{i.}}- \bar{X_{.j}}+\bar{X_{..}}^{2} - (b_{i} - 1)^{2}) \cdot
\sum_{j}(\bar{X_{.j}}-\bar{X_{..}})^{2} \right ]
}

Environmental variance (\code{\link{environmental_variance}})can be expressed as :
\deqn{S^{2}_{xi} = \frac{\sum_{j} (X_{ij} - \bar{X_{i.}})^{2}}{E-1} }

Coefficient of determination may be expressed as:
\deqn{r^{2}_{i} = 1 - \frac{s_{di}^{2}}{s_{xi}^{2}} }

where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{\bar{X_{i.}}} and  \eqn{\bar{X_{.j}}} \cr
denoting marginal means of genotype i and environemnt j,respectively. \cr
\eqn{\bar{X_{..}}} denote the overall mean of X.
}
\examples{
data(Data)
coef.of.determination <- coefficient_of_determination(Data, "Yield", "Genotype", "Environment")
}
\references{
\insertRef{pinthus1973}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{approach}
\keyword{regression}
