% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Wheat APSIM model simulated database}
\format{A dataframe with ? observations on the following 3 variables.
\itemize{
\item{\code{Yield}}{??? }
\item{\code{Genotype}}{genotype, ? varieties}
\item{\code{Environment}}{environment, ? environments}
 }}
\usage{
data(Data)
}
\description{
Multi-environment trail evaluating ? genotypes in ? locations.
}
\examples{
data(Data)
\donttest{boxplot(Yield~Genotype,data=Data,
col='lightblue',border='darkblue')}
}
\references{
\insertRef{casadebaig2016}{toolStability}
}
\keyword{datasets}
