% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Wheat APSIM model simulated database}
\format{A dataframe with 220 observations on the following 3 variables.
\itemize{
\item{\code{Yield}}{kg ha^{-1} }
\item{\code{Genotype}}{genotype, 5 varieties}
\item{\code{Environment}}{environment, 44 environments}
 }}
\usage{
data(Data)
}
\description{
Multi-environment trail evaluating 5 genotypes in 4 locations of 11 years.
}
\examples{
data(Data)
\donttest{boxplot(Yield~Genotype,data=Data,
col='lightblue',border='darkblue')}
}
\references{
\insertRef{casadebaig2016}{toolStability}
}
\keyword{datasets}
