% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_val.R
\name{eco_val}
\alias{eco_val}
\title{Ecovalance}
\usage{
eco_val(data, trait, Genotype, Environment)
}
\arguments{
\item{data}{a dataframe containing at least three column of trait of interest,
genotype, and environment information.}

\item{trait}{the interested variable to be analysize.}

\item{Genotype}{a variable labeling different genotypic varieties}

\item{Environment}{a variable lableing different environmental parameters}
}
\value{
a dataframe with ecovalance indices
}
\description{
\code{eco_val} calculate genetic and environmental interaction.
}
\details{
This function calculate contribution of a genotype to the GE-interactions with dynamic concept, by
squared and summed across all environments for each genotype.

\deqn{S^{2}_{xi} = \frac{\sum_{j}(X_{ij} - \bar{X}_{i.})^2}{E-1}}
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{bar(X)_{i.}} denoting marginal means of genotype i.
}
\examples{
data(Data)
eco_valance <- eco_val(Data,'Yield','Genotype','Environment')
}
\references{
\insertRef{wricke1962}{toolStability}
}
\author{
Tien Cheng Wang (\email{tien.wang@gem.uni-hannover.de})
}
