% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecovalence.R
\name{ecovalence}
\alias{ecovalence}
\title{Ecovalence}
\usage{
ecovalence(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with ecovalence
}
\description{
\code{ecovalence} calculate genetic and environmental interaction.
}
\details{
Ecovalence (Wricke, 1962) is calcualted based on square and sum up the genotype–environment
interaction all over the environment.
Variety with low ecovalence is considered as stable.
Equation of ecovalence can be found in vignette file.
}
\examples{
data(Data)
eco.valance <- ecovalence(Data, "Yield", "Genotype", "Environment")
}
\references{
\insertRef{wricke1962}{toolStability}
}
\author{
Tien-Cheng Wang
}
\keyword{dynamic}
\keyword{stability}
