% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_variance.R
\name{environmental_variance}
\alias{environmental_variance}
\title{Environmental variance}
\usage{
environmental_variance(data, trait, genotype)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}
}
\value{
a data table with environmental variance
}
\description{
\code{environmental_variance} is used to calculate variance of a genotype across environments.
}
\details{
Environmental variance (Roemer, 1917) is calculated by squared and suming up all deviation from genotypic mean for each genotype.
The larger the environmental variance of one genotype is, the lower the stability.

\deqn{S_{xi}^{2} = \frac{\sum_{j} (X_{ij} - \bar{X_{i.}})^{2}}{E-1} }
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{\bar{X_{i.}}} denoting marginal means of genotype i.
}
\examples{
data(Data)
environmental.variance <- environmental_variance(Data,'Yield','Genotype')
}
\references{
\insertRef{roemer1917}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{stability}
\keyword{static}
