% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypic_superiority_measure.R
\name{genotypic_superiority_measure}
\alias{genotypic_superiority_measure}
\title{Genotypic superiority measure}
\usage{
genotypic_superiority_measure(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with genotypic superiority measure
}
\description{
\code{genotypic_superiority_measure} calculate variance of a genotype across environments.
}
\details{
Genotypic superiority measure (Lin and Binns, 1988) is calculatd based on regression function.
Variety with low genotypic superiority measure is considered as stable.

\deqn{P_{i} = \sum_{j}^{n} \frac{(X_{ij}-M_{j})^{2}}{2n}}
where \eqn{X_{ij}} stands for observed trait and
\eqn{M_{j}} stands for maximum response among all culitvars in the jth location.
}
\examples{
data(Data)
res <- genotypic_superiority_measure(Data,'Yield','Genotype','Environment')

}
\references{
\insertRef{lin1988}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{dynamic}
\keyword{stability}
