% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_variance.R
\name{stability_variance}
\alias{stability_variance}
\title{Stability variance}
\usage{
stability_variance(data, trait, genotype, environment)
}
\arguments{
\item{data}{a dataframe containing trait, genotype and environment.}

\item{trait}{colname of a column containing a numeric vector of interested trait to be analysized.}

\item{genotype}{colname of a column containing a character or factor vector labeling different genotypic varieties}

\item{environment}{colname of a column containing a character or factor vector labeling different environments}
}
\value{
a data table with stability variance
}
\description{
\code{stability_variance} calculate variance of a genotype across environments.
}
\details{
Stability variance (Shukla, 1972) is calculatd based on lindear combination of ecovalence and mean square of genotype-environment interaction.
Variety with low stability variance is considered as stable.

\deqn{
\Sigma^{2}_{i} = \frac{1}{(G-1)\cdot(G-2)\cdot(E-1)} \left [
G \cdot(G-1)\cdot \sum_{j} (X_{ij} - \bar{X_{i.}}_\bar{X_{.j}} + \bar{X_{..}})^{2}- \sum_{i} \sum_{j} (X_{ij} - \bar{X_{i.}}_\bar{X_{.j}} + \bar{X_{..}})^{2}
\right ]
}
where \eqn{X_{ij}} is the observed phenotypic mean value of genotype i(i=1,..., G)
in environment j(j=1,...,E), with \eqn{\bar{X_{i.}}} and  \eqn{\bar{X_{.j}}} \cr
denoting marginal means of genotype i and environemnt j,respectively. \cr
\eqn{\bar{X_{..}}} denote the overall mean of X.

Negative values of stability variance is replaced with 0.
}
\examples{
data(Data)
stability.variance <- stability_variance(Data,'Yield','Genotype','Environment')

}
\references{
\insertRef{shukla1972}{toolStability}
}
\author{
Tien Cheng Wang
}
\keyword{dynamic}
\keyword{stability}
